+++ app/src/main/java/com/example/ticketscan/MainActivity.kt
@@ -4,28 +4,27 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.compose.foundation.layout.*
-import androidx.compose.material3.*
-import androidx.compose.runtime.Composable
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
+import androidx.lifecycle.viewmodel.compose.viewModel
+import androidx.navigation.compose.NavHost
+import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
+import com.example.ticketscan.domain.repositories.StatsRepositoryMock
 import com.example.ticketscan.domain.repositories.TicketRepositoryMock
 import com.example.ticketscan.ui.components.TicketScanBottomNavigation
 import com.example.ticketscan.ui.screens.HomeScreen
+import com.example.ticketscan.ui.screens.ProcessingScreen
+import com.example.ticketscan.ui.screens.StatsScreen
+import com.example.ticketscan.ui.screens.StatsViewModel
+import com.example.ticketscan.ui.screens.StatsViewModelFactory
 import com.example.ticketscan.ui.screens.TicketScreen
 import com.example.ticketscan.ui.screens.TicketViewModel
 import com.example.ticketscan.ui.screens.TicketViewModelFactory
 import com.example.ticketscan.ui.theme.TicketScanThemeProvider
-import androidx.lifecycle.viewmodel.compose.viewModel
-import androidx.navigation.compose.NavHost
-import androidx.navigation.compose.composable
-import com.example.ticketscan.domain.repositories.StatsRepositoryMock
-import com.example.ticketscan.ui.screens.StatsScreen
-import com.example.ticketscan.ui.screens.StatsViewModel
-import com.example.ticketscan.ui.screens.StatsViewModelFactory
-import com.example.ticketscan.ui.screens.ProcessingScreen
 import java.util.UUID
 
 class MainActivity : ComponentActivity() {
@@ -78,14 +77,5 @@ class MainActivity : ComponentActivity() {
                 }
             }
         }
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun HomePreview() {
-    TicketScanThemeProvider {
-        val navController = rememberNavController()
-        HomeScreen(navController = navController)
     }
 }
